+++ src/lib.rs
@@ -40,7 +40,7 @@
 //!
 //! const EMPTY_CORE_STATE: ExceptionLock<RefCell<CoreState>> =
 //!     ExceptionLock::new(RefCell::new(CoreState { foo: 0 }));
-//! static CORE_STATE: PerCore<ExceptionLock<RefCell<CoreState>>, CoresImpl, CORE_COUNT> =
+//! static CORE_STATE: PerCore<[ExceptionLock<RefCell<CoreState>>; CORE_COUNT], CoresImpl> =
 //!     PerCore::new([EMPTY_CORE_STATE; CORE_COUNT]);
 //!
 //! fn main() {
@@ -56,6 +56,11 @@
 
 #![no_std]
 
+#[cfg(feature = "alloc")]
+extern crate alloc;
+
+#[cfg(feature = "alloc")]
+mod boxed;
 mod exceptions;
 mod lock;
 
@@ -80,22 +85,26 @@ pub unsafe trait Cores {
 ///
 /// To use this type you must first implement the [`Cores`] trait for your platform.
 ///
-/// `C::core_index()` must always return a value less than `CORE_COUNT` or there will be a runtime
-/// panic.
-pub struct PerCore<T, C: Cores, const CORE_COUNT: usize> {
-    values: [T; CORE_COUNT],
+/// `C::core_index()` must always return a value less than the length of `V` or there will be a
+/// runtime panic.
+#[derive(Default)]
+#[repr(transparent)]
+pub struct PerCore<V: ?Sized, C: Cores> {
     _cores: PhantomData<C>,
+    values: V,
 }
 
-impl<T, C: Cores, const CORE_COUNT: usize> PerCore<T, C, CORE_COUNT> {
+impl<V, C: Cores> PerCore<V, C> {
     /// Creates a new set of per-core values.
-    pub const fn new(values: [T; CORE_COUNT]) -> Self {
+    pub const fn new(values: V) -> Self {
         Self {
             values,
             _cores: PhantomData,
         }
     }
+}
 
+impl<T, C: Cores, const CORE_COUNT: usize> PerCore<[T; CORE_COUNT], C> {
     /// Gets a shared reference to the value for the current CPU core.
     pub fn get(&self) -> &T {
         &self.values[C::core_index()]
@@ -108,6 +117,35 @@ impl<T, C: Cores, const CORE_COUNT: usize> PerCore<T, C, CORE_COUNT> {
 // prevents concurrent access to its contents from different exception contexts. The combination of
 // the two therefore prevents concurrent access to `T`.
 unsafe impl<T: Send, C: Cores, const CORE_COUNT: usize> Sync
-    for PerCore<ExceptionLock<T>, C, CORE_COUNT>
+    for PerCore<[ExceptionLock<T>; CORE_COUNT], C>
 {
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use core::cell::RefCell;
+
+    /// A Fake implementation of `Cores` for test, that will always return 0.
+    pub struct FakeCoresImpl;
+
+    // SAFETY: Tests are all run on a single core.
+    unsafe impl Cores for FakeCoresImpl {
+        fn core_index() -> usize {
+            0
+        }
+    }
+
+    #[test]
+    fn percore_state() {
+        static STATE: PerCore<[ExceptionLock<RefCell<u32>>; 4], FakeCoresImpl> =
+            PerCore::new([const { ExceptionLock::new(RefCell::new(42)) }; 4]);
+
+        {
+            let token = unsafe { ExceptionFree::new() };
+            assert_eq!(*STATE.get().borrow_mut(token), 42);
+            *STATE.get().borrow_mut(token) += 1;
+            assert_eq!(*STATE.get().borrow_mut(token), 43);
+        }
+    }
 }